Changed architecture with IMovieRatingService and MovieRatingService.
