+++ MovieRating/MovieRating.Core/MovieRating.Core.csproj
@@ -4,4 +4,8 @@
     <TargetFramework>netstandard2.0</TargetFramework>
   </PropertyGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\MovieRating.Infrastructure.Data\MovieRating.Infrastructure.Data.csproj" />
+  </ItemGroup>
+
 </Project>
