+++ MovieRating/MovieRating.Infrastructure.Data/MovieRating.Infrastructure.Data.csproj
@@ -10,7 +10,6 @@
 
   <ItemGroup>
     <ProjectReference Include="..\MovieRating.Core.Entities\MovieRating.Core.Entities.csproj" />
-    <ProjectReference Include="..\MovieRating.Core\MovieRating.Core.csproj" />
   </ItemGroup>
 
 </Project>
